requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
contourpy==1.2.0
cycler==0.12.1
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.14.0
docutils==0.20.1
flatbuffers==23.5.26
fonttools==4.49.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
gTTS==2.5.1
gunicorn==21.2.0
h5py==3.10.0
idna==3.6
jax==0.4.25
keras==2.15.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mediapipe==0.10.10
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
protobuf==3.20.3
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.12.0
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.4
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
whitenoise==6.6.0
wrapt==1.14.1